+++ tests/src/verification_tests/introspection_tests.rs
@@ -75,6 +75,7 @@ async fn verification_test_setup() -> VerificationTestContext {
         ix_metadata_pointer: None,
         ix_metadata: None,
         ix_scaled_ui_amount: None,
+        ix_default_account_state: None,
     };
 
     initialize_mint(
@@ -443,6 +444,7 @@ async fn test_update_metadata_under_verification() {
             additional_metadata: vec![],
         }),
         ix_scaled_ui_amount: None,
+        ix_default_account_state: None,
     };
 
     initialize_mint(